vsc_generator: remove code related to template engine (jinja2)
An attempt to reverse dependencies by moving all the jinja2 related code
to the vsc.templates module. This makes vsc_generator.py more abstract
and in theory removes dependency on the concrete template engine.
