+++ vsc/model/vsc_generator.py
@@ -18,31 +18,8 @@ VSC code-generation functions
 #    according to given templates.
 
 # For other features from parser module
-from vsc.model import vsc_ast
-from vsc.templates import TemplatePath
-from typing import Generic, Any
-import jinja2
-import sys
-
-# Set up Jinja
-jinja_env = jinja2.Environment(
-        # Use the subdirectory 'templates' relative to this file's location
-        loader=jinja2.FileSystemLoader(TemplatePath),
-
-        # Templates with these extension gets automatic auto escape for HTML
-        # It's more annoying for code generation, so passing empty list for now.
-        autoescape=jinja2.select_autoescape([])
-        )
-
-# We want the control blocks in the template to NOT result in any extra
-# white space when rendering templates. However, this might be a choice
-# made by each generator, so we need to export the ability to keep these
-# settings public for other code to modify them.
-jinja_env.trim_blocks = True
-jinja_env.lstrip_blocks = True
-jinja_env.undefined = jinja2.StrictUndefined
-
-default_templates = {}
+from typing import Any
+from vsc.templates import set_template_env, get_template, get_default_template, render_template
 
 # Exception:
 class GeneratorError(BaseException):
@@ -50,11 +27,6 @@ class GeneratorError(BaseException):
         self.msg = m
 
 # ---------- GENERATION FUNCTIONS ------------
-
-
-# Get template with given name (search path should be handled by the loader)
-def get_template(filename):
-    return jinja_env.get_template(filename)
 
 # gen() function to be called from templates
 def gen(node : Any, template_file = None):
@@ -100,7 +72,7 @@ def _gen_with_default_template(node : Any):
         return node
 
     # Complex types -> use the corresponding template for the type
-    tpl = default_templates.get(nodetype)
+    tpl = get_default_template(nodetype)
     if tpl is None:
         raise GeneratorError(f"gen() function called with node of type '{nodetype}' but no default template is defined for this type.")
     else:
@@ -117,7 +89,7 @@ def gen_template_text(node: Any, template_text: str):
    if template_text is None:
        raise GeneratorError(f'gen_template_text called without template')
    elif type(template_text) == str:
-       return jinja2.Template(template_text).render({'item' : node})
+       return render_template(template_text,{'item' : node})
    else:
        print(f'node is of type {type(node)}, second arg is of type {type(template_text)}  ({type(template_text).__class__}, {type(template_text).__class__.__name__})')
        raise GeneratorError(f'Wrong use of gen() function! Usage: pass the node as first argument (you passed a {type(node)}), and optionally template name (str) as second argument. (You passed a {template_text.__name__})')
@@ -128,30 +100,4 @@ def gen_dict_with_template_file(variables : dict, templatefile):
 
 # Export the gen() function and classes into jinja template land
 # so that they can be referred to inside templates.
-jinja_env.globals.update(gen=gen)
-
-
-def usage():
-    print("usage: generator.py <input.yaml-file (path)> <output-template-file (name only, not path)>")
-    sys.exit(1)
-
-# This is a default definition for our current generation tests.
-# It may need to be changed, or defined differently in a custom generator
-default_templates = {
-        'AST': 'AST-simple_doc.tpl',
-        'Service': 'Service-simple_doc.html',
-        'Argument': 'Argument-simple_doc.html',
-        'Error': 'Error-simple_doc.html',
-        'Member': 'Member-simple_doc.html',
-        'Namespace': 'Namespace-simple_doc.html',
-        'Argument': 'Argument-simple_doc.html'
-        }
-
-# If run as a script, generate a single YAML file using a single root template
-if __name__ == "__main__":
-    if not len(sys.argv) == 3:
-        usage()
-    ast = vsc_ast.read_ast_from_yaml_file(sys.argv[1])
-    templatename = sys.argv[2]
-    print(gen(ast, templatename))
-
+set_template_env(gen=gen)
